Refactor syncstore.js to use generic loadable backends.
The original kvstore-backend syncstore implementation is now in
lib/syncstore/kvstore.js.  Other dedicated backends to follow.

The notification-endpoint code has been factored out so that we don't
have to duplicate it across backend implementations.
