+++ lib/config.js
@@ -5,7 +5,8 @@
 const fs = require('fs');
 const convict = require('convict');
 
-const AVAILABLE_BACKENDS = ["memory", "mysql"];
+const KVSTORE_AVAILABLE_BACKENDS = ["memory", "mysql"];
+const SYNCSTORE_AVAILABLE_BACKENDS = ["kvstore"];
 
 
 var conf = module.exports = convict({
@@ -34,13 +35,24 @@ var conf = module.exports = convict({
   },
   kvstore: {
     backend: {
-      format: AVAILABLE_BACKENDS,
+      format: KVSTORE_AVAILABLE_BACKENDS,
       default: "memory",
       env: 'KVSTORE_BACKEND'
     },
     available_backends: {
       doc: "List of available key-value stores",
-      default: AVAILABLE_BACKENDS
+      default: KVSTORE_AVAILABLE_BACKENDS
+    }
+  },
+  syncstore: {
+    backend: {
+      format: SYNCSTORE_AVAILABLE_BACKENDS,
+      default: "kvstore",
+      env: 'SYNCSTORE_BACKEND'
+    },
+    available_backends: {
+      doc: "List of available syncstore implementations",
+      default: SYNCSTORE_AVAILABLE_BACKENDS
     }
   },
   mysql: {
